refactor(ui): rename first-party-fields to components
Rename the robot schema form's first-party-fields directory to components
and update the imports in schema-form.tsx. Also clean up leftovers from
removing the 'Show default fields' toggle: drop the now-unused useState
and Switch imports and update the two schema-form tests that toggled the
removed switch.
