+++ application/ui/src/features/robots/robot-form/robot-schema/components/info-field.tsx
